Move test helpers to be layer specific and refine vi.mock targets
