+++ tests/repositories/coding.test.ts
@@ -1,4 +1,4 @@
-import { mockDb } from '../database.helper.ts';
+import { mockDb } from './repository.helper.ts';
 import { BaseRepository } from '../../src/repositories/base.ts';
 import { CodingRepository } from '../../src/repositories/coding.ts';
